backup: Move all Snapshot creation fields out of kcontext.
* Removes everything used only in snapshot creation (backup/sync/ptar)
etc to a dedicated BuilderOptions type like we do elsewhere.

* This is the easy part of the diff, plakar is another beast and will
come soon after.

* Note that apart from snap header the Hostname is also used by Locks()
(as an hint to the user), we tap into the Header to retrieve this info
as a fallback, since Locks needs to go away anyway.

* Note2: I'm getting rid of kcontext_test.go, sorry sayoun@ but kcontext
is meant to die and the tests are just introducing churn for the sake
of churn in this particular instance.