backup: fix handling of excludes
When doing the pooling I forgot to turn that `return' into a
`continue', so instead of skipping forgotten files, we quit one worker
when we find an ignored files.

The other scanresults are eaten away in the draining loop that's
intended to be used for cancellation.

Adding a regress while here.