Packer: Change the interface a bit.
* Rather than having the flusher (PutPackfile) do the file format setup,
provide a single Serialize() point and let individual Packfile
implementation just Serialize the data in the expected (and documented
format).

* Note that this introduces a slight behavior change (a fix actually),
for SerializeFooter was _reincluding_  the index, and we got two indexes
for the price of one. Switching to the (at the time) unused Serialize
removes that double index.

* Introduce a Cleanup function, disk based packfiles will have some
cleanup to do in the close path.