snapshot: Return all packfiles per blob in ListPackfiles().
* ListPackfiles() is used by the maintenance to decide if certain blobs
  (and moreoever the packfiles containing them) are still reachable
(think deleted snapshots). The problem is that by default our state
returns _the_ first location (packfile+offset) that matches a Blob mac.
This is fine for most of operations, but for the purpose of maintenance
that was not enough, as we ended up deleting all packfiles that are not
the first in the state (that was made way more evident by our sequential
packer).