repository: Fix over prefetching.
* When prefetching an object content, respect the maxSize parameter.

* The problem lied in the fact that we keep track of the current
  "segment" size through the `size` variable, but this gets reset on the
next segment construction. Since the VFS prefetcher uses 20mb as maxSize
and our packfiles are around 20MB we were always crossing a boundary
therefore resetting that `size` variable hence never exiting the
iteration. Meaning that the prefetch buffer would grow up to the size of
the file.

* Fix this by introducing an accumulatedSize that tracks well, the
  actual number of bytes we have been giving to the caller and that
stops when we are above the threshold.